refactor DriverManager.close(driver)
